requirements.txt
fastapi>=0.100.0
uvicorn>=0.23.0
sqlalchemy>=2.0.0
python-multipart>=0.0.6
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
pydantic-settings>=2.0.0
numpy>=1.24.0
opencv-python>=4.8.0
face_recognition>=1.3.0
jinja2>=3.1.0
aiofiles>=23.0.0
requests>=2.31.0
